+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Refactor `AgeIn` helper class to replace `ageInSeconds()`, `ageInHours()`, etc. and export it
 
 ### v1.1.17
 * Add `debugWithTraceObjs()` method to `Logger` class